requirements.txt
requests>=2.31.0
websockets>=11.0
Pillow>=9.0.0
svglib>=1.5.0
reportlab>=3.6.0
lxml>=4.9.0
cairosvg>=2.7.0
inky[rpi,fonts]>=1.4.0

# Note: cairosvg requires system Cairo library to be installed first:
#   On Linux (Raspberry Pi): sudo apt-get install libcairo2-dev
#   On macOS: brew install cairo
#   Then: pip install -r requirements.txt
#
# Note: inky[rpi,fonts] is for Pimoroni Inky Impression displays
#   On Raspberry Pi, this should work out of the box after pip install
